backend/requirements.txt
fastapi
yfinance
google-generativeai
python-dotenv
uvicorn
